LayoutTests/imported/w3c/web-platform-tests/css/css-text-decor/parsing/text-decoration-line-valid-expected.txt


PASS e.style['text-decoration-line'] = "none" should set the property value
PASS e.style['text-decoration-line'] = "underline" should set the property value
PASS e.style['text-decoration-line'] = "overline" should set the property value
PASS e.style['text-decoration-line'] = "line-through" should set the property value
PASS e.style['text-decoration-line'] = "blink" should set the property value
PASS e.style['text-decoration-line'] = "underline overline" should set the property value
FAIL e.style['text-decoration-line'] = "overline underline" should set the property value assert_equals: serialization should be canonical expected "underline overline" but got "overline underline"
PASS e.style['text-decoration-line'] = "underline line-through" should set the property value
FAIL e.style['text-decoration-line'] = "line-through underline" should set the property value assert_equals: serialization should be canonical expected "underline line-through" but got "line-through underline"
PASS e.style['text-decoration-line'] = "underline blink" should set the property value
FAIL e.style['text-decoration-line'] = "blink underline" should set the property value assert_equals: serialization should be canonical expected "underline blink" but got "blink underline"
PASS e.style['text-decoration-line'] = "overline line-through" should set the property value
FAIL e.style['text-decoration-line'] = "line-through overline" should set the property value assert_equals: serialization should be canonical expected "overline line-through" but got "line-through overline"
PASS e.style['text-decoration-line'] = "overline blink" should set the property value
FAIL e.style['text-decoration-line'] = "blink overline" should set the property value assert_equals: serialization should be canonical expected "overline blink" but got "blink overline"
PASS e.style['text-decoration-line'] = "line-through blink" should set the property value
FAIL e.style['text-decoration-line'] = "blink line-through" should set the property value assert_equals: serialization should be canonical expected "line-through blink" but got "blink line-through"
PASS e.style['text-decoration-line'] = "underline overline line-through" should set the property value
FAIL e.style['text-decoration-line'] = "underline line-through overline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through" but got "underline line-through overline"
FAIL e.style['text-decoration-line'] = "overline underline line-through" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through" but got "overline underline line-through"
FAIL e.style['text-decoration-line'] = "overline line-through underline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through" but got "overline line-through underline"
FAIL e.style['text-decoration-line'] = "line-through underline overline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through" but got "line-through underline overline"
FAIL e.style['text-decoration-line'] = "line-through overline underline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through" but got "line-through overline underline"
PASS e.style['text-decoration-line'] = "underline overline blink" should set the property value
FAIL e.style['text-decoration-line'] = "underline blink overline" should set the property value assert_equals: serialization should be canonical expected "underline overline blink" but got "underline blink overline"
FAIL e.style['text-decoration-line'] = "overline underline blink" should set the property value assert_equals: serialization should be canonical expected "underline overline blink" but got "overline underline blink"
FAIL e.style['text-decoration-line'] = "overline blink underline" should set the property value assert_equals: serialization should be canonical expected "underline overline blink" but got "overline blink underline"
FAIL e.style['text-decoration-line'] = "blink underline overline" should set the property value assert_equals: serialization should be canonical expected "underline overline blink" but got "blink underline overline"
FAIL e.style['text-decoration-line'] = "blink overline underline" should set the property value assert_equals: serialization should be canonical expected "underline overline blink" but got "blink overline underline"
PASS e.style['text-decoration-line'] = "underline line-through blink" should set the property value
FAIL e.style['text-decoration-line'] = "underline blink line-through" should set the property value assert_equals: serialization should be canonical expected "underline line-through blink" but got "underline blink line-through"
FAIL e.style['text-decoration-line'] = "line-through underline blink" should set the property value assert_equals: serialization should be canonical expected "underline line-through blink" but got "line-through underline blink"
FAIL e.style['text-decoration-line'] = "line-through blink underline" should set the property value assert_equals: serialization should be canonical expected "underline line-through blink" but got "line-through blink underline"
FAIL e.style['text-decoration-line'] = "blink underline line-through" should set the property value assert_equals: serialization should be canonical expected "underline line-through blink" but got "blink underline line-through"
FAIL e.style['text-decoration-line'] = "blink line-through underline" should set the property value assert_equals: serialization should be canonical expected "underline line-through blink" but got "blink line-through underline"
PASS e.style['text-decoration-line'] = "overline line-through blink" should set the property value
FAIL e.style['text-decoration-line'] = "overline blink line-through" should set the property value assert_equals: serialization should be canonical expected "overline line-through blink" but got "overline blink line-through"
FAIL e.style['text-decoration-line'] = "line-through overline blink" should set the property value assert_equals: serialization should be canonical expected "overline line-through blink" but got "line-through overline blink"
FAIL e.style['text-decoration-line'] = "line-through blink overline" should set the property value assert_equals: serialization should be canonical expected "overline line-through blink" but got "line-through blink overline"
FAIL e.style['text-decoration-line'] = "blink overline line-through" should set the property value assert_equals: serialization should be canonical expected "overline line-through blink" but got "blink overline line-through"
FAIL e.style['text-decoration-line'] = "blink line-through overline" should set the property value assert_equals: serialization should be canonical expected "overline line-through blink" but got "blink line-through overline"
PASS e.style['text-decoration-line'] = "underline overline line-through blink" should set the property value
FAIL e.style['text-decoration-line'] = "underline overline blink line-through" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "underline overline blink line-through"
FAIL e.style['text-decoration-line'] = "underline line-through overline blink" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "underline line-through overline blink"
FAIL e.style['text-decoration-line'] = "underline line-through blink overline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "underline line-through blink overline"
FAIL e.style['text-decoration-line'] = "underline blink overline line-through" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "underline blink overline line-through"
FAIL e.style['text-decoration-line'] = "underline blink line-through overline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "underline blink line-through overline"
FAIL e.style['text-decoration-line'] = "overline underline line-through blink" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "overline underline line-through blink"
FAIL e.style['text-decoration-line'] = "overline underline blink line-through" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "overline underline blink line-through"
FAIL e.style['text-decoration-line'] = "overline line-through underline blink" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "overline line-through underline blink"
FAIL e.style['text-decoration-line'] = "overline line-through blink underline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "overline line-through blink underline"
FAIL e.style['text-decoration-line'] = "overline blink underline line-through" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "overline blink underline line-through"
FAIL e.style['text-decoration-line'] = "overline blink line-through underline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "overline blink line-through underline"
FAIL e.style['text-decoration-line'] = "line-through underline overline blink" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "line-through underline overline blink"
FAIL e.style['text-decoration-line'] = "line-through underline blink overline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "line-through underline blink overline"
FAIL e.style['text-decoration-line'] = "line-through overline underline blink" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "line-through overline underline blink"
FAIL e.style['text-decoration-line'] = "line-through overline blink underline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "line-through overline blink underline"
FAIL e.style['text-decoration-line'] = "line-through blink underline overline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "line-through blink underline overline"
FAIL e.style['text-decoration-line'] = "line-through blink overline underline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "line-through blink overline underline"
FAIL e.style['text-decoration-line'] = "blink underline overline line-through" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "blink underline overline line-through"
FAIL e.style['text-decoration-line'] = "blink underline line-through overline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "blink underline line-through overline"
FAIL e.style['text-decoration-line'] = "blink overline underline line-through" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "blink overline underline line-through"
FAIL e.style['text-decoration-line'] = "blink overline line-through underline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "blink overline line-through underline"
FAIL e.style['text-decoration-line'] = "blink line-through underline overline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "blink line-through underline overline"
FAIL e.style['text-decoration-line'] = "blink line-through overline underline" should set the property value assert_equals: serialization should be canonical expected "underline overline line-through blink" but got "blink line-through overline underline"
FAIL e.style['text-decoration-line'] = "spelling-error" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['text-decoration-line'] = "grammar-error" should set the property value assert_not_equals: property should be set got disallowed value ""
